# docs/swagger.yaml
basePath: /api
definitions:
  database.Chirp:
    properties:
      body:
        type: string
      created_at:
        type: string
      id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  main.chirpResponse:
    properties:
      body:
        type: string
      created_at:
        type: string
      id:
        type: string
      token:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  main.createChirpParameters:
    properties:
      body:
        type: string
    type: object
  main.createUserParameters:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  main.createUserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_chirpy_red:
        type: boolean
      updated_at:
        type: string
    type: object
  main.errorResponse:
    properties:
      error:
        type: string
    type: object
  main.loginParameters:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  main.loginResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_chirpy_red:
        type: boolean
      refresh_token:
        type: string
      token:
        type: string
      updated_at:
        type: string
    type: object
  main.updateUserParameters:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  main.updateUserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_chirpy_red:
        type: boolean
      updated_at:
        type: string
    type: object
  main.webhookParameters:
    properties:
      data:
        properties:
          user_id:
            type: string
        type: object
      event:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API for Chirpy.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Chirpy API
  version: "1.0"
paths:
  /api/polka/webhooks:
    post:
      consumes:
      - application/json
      description: Handle webhook from Polka to upgrade user to Chirpy Red
      parameters:
      - description: API Key
        in: header
        name: Authorization
        required: true
        type: string
      - description: Webhook parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.webhookParameters'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Upgrade user to Chirpy Red
      tags:
      - webhooks
  /chirps:
    get:
      consumes:
      - application/json
      description: Get a list of chirps, optionally filtered by author ID
      parameters:
      - description: Filter by author ID
        in: query
        name: author_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Chirp'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: List chirps
      tags:
      - chirps
    post:
      consumes:
      - application/json
      description: Create a new chirp with the given body
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Chirp parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.createChirpParameters'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.chirpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Create a chirp
      tags:
      - chirps
  /login:
    post:
      consumes:
      - application/json
      description: Login with email and password to get access and refresh tokens
      parameters:
      - description: Login credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.loginParameters'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.loginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Login a user
      tags:
      - auth
  /users:
    post:
      consumes:
      - application/json
      description: Create a new user with email and password
      parameters:
      - description: User parameters
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.createUserParameters'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.createUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Create a user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user email and password
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User parameters
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.updateUserParameters'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.updateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Update a user
      tags:
      - users
schemes:
- http
- https
swagger: "2.0"